Backend/ocr-module/sign/devops/sign/app/utils/signature_utils.py
from PIL import Image
import io

# Dummy function to simulate signature detection
def detect_signature(image_file):
    # In a real app, you would use a deep learning model or algorithm here
    img = Image.open(image_file)
    return img

# Dummy function to simulate signature comparison
def compare_signatures(uploaded_signature, saved_signature):
    # Here you can use an actual comparison method (e.g., cosine similarity, deep learning)
    return True  # Just returning True for simulation purposes
