Backend/ocr-module/sign/devops/sign/requirements.txt
flask
flask_sqlalchemy
Pillow
opencv-python
tensorflow
